ebdmaxcloud/requirements.txt
Flask==2.2.2
requests==2.28.1
python-dotenv==0.21.0
pytz==2023.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Chart.js==3.9.1